Move TungstenCache code into own file; avoid unnecesary row format conversions.
