+++ sql/core/src/main/scala/org/apache/spark/sql/SQLImplicits.scala
@@ -17,27 +17,16 @@
 
 package org.apache.spark.sql
 
-import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
-
-import scala.collection.mutable.ArrayBuffer
 import scala.language.implicitConversions
 import scala.reflect.runtime.universe.TypeTag
 
-import com.google.common.io.ByteStreams
-
-import org.apache.spark.SparkEnv
 import org.apache.spark.annotation.Experimental
 import org.apache.spark.io.CompressionCodec
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.catalyst.{CatalystTypeConverters, InternalRow}
-import org.apache.spark.sql.catalyst.expressions.{SpecificMutableRow, UnsafeProjection, UnsafeRow}
-import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.sources.{BaseRelation, TableScan}
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.catalyst.expressions.SpecificMutableRow
+import org.apache.spark.sql.execution.TungstenCache
 import org.apache.spark.sql.types._
-import org.apache.spark.storage.StorageLevel
-import org.apache.spark.unsafe.Platform
-import org.apache.spark.unsafe.array.ByteArrayMethods
-import org.apache.spark.unsafe.memory.{MemoryBlock, TaskMemoryManager}
 import org.apache.spark.unsafe.types.UTF8String
 
 /**
@@ -127,167 +116,24 @@ private[sql] abstract class SQLImplicits {
   /**
    * ::Experimental::
    *
-   * Pimp my library decorator for tungsten caching of DataFrames.
-   * @since 1.5.1
+   * Pimp my library decorator for Tungsten caching of DataFrames.
+   * @since 1.6.0
    */
   @Experimental
-  implicit class TungstenCache(df: DataFrame) {
+  implicit class TungstenCacheImplicits(df: DataFrame) {
     /**
      * Packs the rows of [[df]] into contiguous blocks of memory.
      * @param compressionType "" (default), "lz4", "lzf", or "snappy", see
      *                        [[CompressionCodec.ALL_COMPRESSION_CODECS]]
      * @param blockSize size of each MemoryBlock (default = 4 MB)
      */
-    def tungstenCache(
-        compressionType: String = "",
-        blockSize: Int = 4000000): (RDD[_], DataFrame) = {
-      val schema = df.schema
-
-      val convert = CatalystTypeConverters.createToCatalystConverter(schema)
-      val internalRows = df.rdd.map(convert(_).asInstanceOf[InternalRow])
-      val cachedRDD = internalRows.mapPartitions { rowIterator =>
-        val bufferedRowIterator = rowIterator.buffered
-        val convertToUnsafe = UnsafeProjection.create(schema)
-        val taskMemoryManager = new TaskMemoryManager(SparkEnv.get.executorMemoryManager)
-        val compressionCodec: Option[CompressionCodec] = if (compressionType.isEmpty) {
-          None
-        } else {
-          Some(CompressionCodec.createCodec(SparkEnv.get.conf, compressionType))
-        }
-
-        new Iterator[MemoryBlock] {
-          // NOTE: This assumes that size of every row < blockSize
-          def next(): MemoryBlock = {
-            // Packs rows into a `blockSize` bytes contiguous block of memory, starting a new block
-            // whenever the current fills up
-            // Each row is laid out in memory as [rowSize (4)|row (rowSize)]
-            val block = taskMemoryManager.allocateUnchecked(blockSize)
-
-            var currOffset = 0
-            val seenRows = ArrayBuffer[InternalRow]()
-            while (bufferedRowIterator.hasNext && currOffset < blockSize) {
-              val currRow = convertToUnsafe.apply(bufferedRowIterator.head)
-              val recordSize = 4 + currRow.getSizeInBytes
-              if (currOffset + recordSize < blockSize) {
-                Platform.putInt(
-                  block.getBaseObject, block.getBaseOffset + currOffset, currRow.getSizeInBytes)
-                currRow.writeToMemory(block.getBaseObject, block.getBaseOffset + currOffset + 4)
-                seenRows += currRow
-                bufferedRowIterator.next()
-              }
-              currOffset += recordSize // Increment currOffset regardless to break loop when full
-            }
-
-            // Optionally compress block before writing
-            compressionCodec match {
-              case Some(codec) =>
-                // Compress the block using an on-heap byte array
-                val compressedBlockArray: Array[Byte] = {
-                  val blockArray = new Array[Byte](blockSize)
-                  Platform.copyMemory(
-                    block.getBaseObject,
-                    block.getBaseOffset,
-                    blockArray,
-                    Platform.BYTE_ARRAY_OFFSET,
-                    blockSize)
-                  val baos = new ByteArrayOutputStream(blockSize)
-                  val compressedBaos = codec.compressedOutputStream(baos)
-                  compressedBaos.write(blockArray)
-                  compressedBaos.flush()
-                  compressedBaos.close()
-                  baos.toByteArray
-                }
-
-                // Allocate a new block with compressed byte array padded to word boundary
-                val totalRecordSize = compressedBlockArray.length + 4
-                val nearestWordBoundary =
-                  ByteArrayMethods.roundNumberOfBytesToNearestWord(totalRecordSize)
-                val padding = nearestWordBoundary - totalRecordSize
-                val compressedBlock = taskMemoryManager.allocateUnchecked(totalRecordSize + padding)
-                Platform.putInt(
-                  compressedBlock.getBaseObject,
-                  compressedBlock.getBaseOffset,
-                  padding)
-                Platform.copyMemory(
-                  compressedBlockArray,
-                  Platform.BYTE_ARRAY_OFFSET,
-                  compressedBlock.getBaseObject,
-                  compressedBlock.getBaseOffset + 4,
-                  compressedBlockArray.length)
-                taskMemoryManager.freeUnchecked(block)
-                compressedBlock
-              case None => block
-            }
-          }
-
-          def hasNext: Boolean = bufferedRowIterator.hasNext
-        }
-      }.setName(compressionType + "_" + df.toString).persist(StorageLevel.MEMORY_ONLY)
-
-      val baseRelation: BaseRelation = new BaseRelation with TableScan {
-        override val sqlContext = _sqlContext
-        override val schema = df.schema
-        override val needConversion = false
-
-        override def buildScan(): RDD[Row] = {
-          val numFields = this.schema.length
-          val _compressionType: String = compressionType
-          val _blockSize = blockSize
-
-          cachedRDD.flatMap { rawBlock =>
-            // Optionally decompress block
-            val compressionCodec: Option[CompressionCodec] = if (_compressionType.isEmpty) {
-              None
-            } else {
-              Some(CompressionCodec.createCodec(SparkEnv.get.conf, _compressionType))
-            }
-            val block = compressionCodec match {
-              case Some(codec) =>
-                // Copy compressed block (excluding padding) to on-heap byte array
-                val padding = Platform.getInt(rawBlock.getBaseObject, rawBlock.getBaseOffset)
-                val compressedBlockArray = new Array[Byte](_blockSize)
-                Platform.copyMemory(
-                  rawBlock.getBaseObject,
-                  rawBlock.getBaseOffset + 4,
-                  compressedBlockArray,
-                  Platform.BYTE_ARRAY_OFFSET,
-                  rawBlock.size() - padding)
-
-                // Decompress into MemoryBlock backed by on-heap byte array
-                val decompressionStream =
-                  codec.compressedInputStream(new ByteArrayInputStream(compressedBlockArray))
-                val decompressedBlock = new Array[Byte](_blockSize)
-                ByteStreams.readFully(decompressionStream, decompressedBlock)
-                decompressionStream.close()
-                MemoryBlock.fromByteArray(decompressedBlock)
-              case None => rawBlock
-            }
-
-            val rows = new ArrayBuffer[InternalRow]()
-            var currOffset = 0
-            var moreData = true
-            while (currOffset < block.size() && moreData) {
-              val rowSize = Platform.getInt(block.getBaseObject, block.getBaseOffset + currOffset)
-              currOffset += 4
-              // TODO: should probably have a null terminator rather than relying on zeroed out
-              if (rowSize > 0) {
-                val unsafeRow = new UnsafeRow()
-                unsafeRow.pointTo(
-                  block.getBaseObject, block.getBaseOffset + currOffset, numFields, rowSize)
-                rows.append(unsafeRow)
-                currOffset += rowSize
-              } else {
-                moreData = false
-              }
-            }
-            rows
-          }.asInstanceOf[RDD[Row]]
-        }
-
-        override def toString: String = getClass.getSimpleName + s"[${df.toString}]"
-      }
-      cachedRDD.count // trigger caching action because it's lazy
-      (cachedRDD, DataFrame(_sqlContext, LogicalRelation(baseRelation)))
+    def tungstenCache(compressionType: String = "", blockSize: Int = 4000000): DataFrame = {
+      val cached: RDD[InternalRow] =
+        TungstenCache.cache(
+          df.queryExecution.sparkPlan,
+          if (compressionType.isEmpty) None else Some(compressionType),
+          blockSize)
+      _sqlContext.internalCreateDataFrame(cached, df.schema)
     }
   }
 }
